requirements-dev.txt
# requirements-dev.txt

# Include runtime dependencies needed for tests and development
-r requirements.txt

# Development and Testing Tools (pin versions for reproducibility)
pytest==8.3.5
pytest-cov==5.0.0
mypy==1.10.0 # Or your desired Mypy version
ruff==0.4.4  # Or your desired Ruff version
bandit==1.7.9
safety==3.5.1
pre-commit==3.7.1 # Or latest stable
types-PyYAML==6.0.12.20240311 # Stubs for PyYAML (if used by any tool or MyPy hook)
pandas-stubs==2.2.2.240514   # Stubs for Pandas